requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0
langchain-openai>=0.2.0
langchain-mistralai>=1.0.1
langgraph>=1.0.0
langsmith[otel]>=0.4.26
composio>=0.1.0
python-dotenv>=1.0.1
schedule>=1.2.0
fastapi>=0.121.0
uvicorn>=0.38.0
python-multipart>=0.0.9
apscheduler>=3.10.0
langserve>=0.3.0
sse-starlette>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
huggingface-hub>=0.16.4
Pillow>=10.0.0
astrapy>=2.0,<3.0
